+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file, following t
 
 Note that since we don't clearly distinguish between a public and private interfaces there will be changes in non-major versions that are potentially breaking. If we make breaking changes to less used interfaces we will highlight it in here.
 
+## [Unreleased]
+
+- Fix BinaryCIF encoder edge cases caused by re-encoding an existing BinaryCIF file
+
 
 ## [v4.0.1] - 2023-02-19